fixup! [fuzz] Add fuzzutil package and fuzz.go
